PDKS_Panel/WebService/requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
psycopg2-binary==2.9.9
python-dotenv==0.19.0
Werkzeug==2.0.1
click==8.0.1
colorama==0.4.4
pyzk==0.9
pandas==2.1.4
reportlab==4.1.0
openpyxl==3.1.2
xlsxwriter==3.1.9
Pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1
Flask-WTF==1.2.1
WTForms==3.1.2
requests==2.31.0
urllib3==2.2.1
Jinja2==3.1.3
itsdangerous==2.1.2
MarkupSafe==2.1.5
six==1.16.0
typing_extensions==4.9.0
numpy==1.26.4
tzdata==2024.1
SQLAlchemy==1.4.41
Flask-SQLAlchemy==2.5.1
pdfkit==1.0.0 